feat(docker-compose.yml): add docker-compose file to set up PostgreSQL database container docs(readme.md): add explanation of docker-compose commands and configuration options
The docker-compose.yml file is added to the project to set up a PostgreSQL database container. The version of the compose file is set to "3.1". The services section defines the services to be run.

The "db" service is defined with the image "postgres:14.2-alpine". It has the restart policy set to "always" to ensure that the container restarts automatically if it stops. The environment variable "POSTGRES_PASSWORD" is set to "example" to set the password for the PostgreSQL database. The container's port 5432 is mapped to the host's port 5432 to allow access to the database.

The readme.md file is updated to provide an explanation of the docker-compose commands and configuration options used in the docker-compose.yml file. It also mentions that the POSTGRES_USER and POSTGRES_DB variables are optional and will use default values if not set. The source for more information about the PostgreSQL image and its options is provided as a link. The example uses the Alpine version of the image for its lightweight nature.