feat(migrations): add migration to create transactions table
This commit adds a new migration file `20231106155831_create_table_transactions.js` that creates a `transactions` table in the database. The table has columns for `id`, `description`, `type`, `date`, `amount`, `status`, and `acc_idd`. The `acc_idd` column is a foreign key referencing the `id` column of the `accounts` table.

feat(node.migrate.js): add script to apply migrations

This commit adds a new file `node.migrate.js` that contains a script to apply the database migrations using the `app.db.migrate.latest()` function. If the migrations are applied successfully, it logs a success message and exits with code 0. If there is an error applying the migrations, it logs an error message and exits with code 1.